+++ src/App.tsx
@@ -5,6 +5,7 @@ import { CompetitorIntelligence } from './pages/CompetitorIntelligence';
 import { DataExplorer } from './pages/DataExplorer';
 import { InsightCenter } from './pages/InsightCenter';
 import { MarketLandscape } from './pages/MarketLandscape';
+import { MoleculeExplorer } from './pages/MoleculeExplorer';
 import { Methodology } from './pages/Methodology';
 import { OpportunitySignals } from './pages/OpportunitySignals';
 import { Overview } from './pages/Overview';
@@ -17,6 +18,8 @@ function Router() {
       return <MarketLandscape />;
     case 'brand':
       return <BrandPerformance />;
+    case 'molecules':
+      return <MoleculeExplorer />;
     case 'competitors':
       return <CompetitorIntelligence />;
     case 'opportunities':